+++ modules/core/src/main/scala/com/odenzo/ripple/bincodec/codecs/ContainerFields.scala
@@ -9,7 +9,7 @@ import spire.math.UByte
 import com.odenzo.ripple.bincodec._
 import com.odenzo.ripple.bincodec.decoding.TxBlobBuster
 import com.odenzo.ripple.bincodec.encoding.{TypeSerializers, CodecUtils}
-import com.odenzo.ripple.bincodec.reference.{FieldData, FieldMetaData}
+import com.odenzo.ripple.bincodec.reference.{FieldData, FieldMetaData, DefinitionData}
 import com.odenzo.ripple.bincodec.utils.JsonUtils
 
 trait STObjectCodec extends CodecUtils with JsonUtils {
@@ -18,28 +18,28 @@ trait STObjectCodec extends CodecUtils with JsonUtils {
     * Top level object has no matching FieldInfo :-/
     *
     * @param o
-    * @param isNested  If a nested object somethings are serialized differenty.
     * @param isSigning If true only signing fields serialized, or all serializable
     *                  This is determines from definitions.json
     *
     * @return List of each of the named fields in the object with their encoded information
     */
-  def encodeSTObject(o: Json, isNested: Boolean, isSigning: Boolean): Either[BinCodecLibError, EncodedSTObject] = {
+  def encodeSTObject(o: Json, isSigning: Boolean): Either[BinCodecLibError, EncodedSTObject] = {
+    scribe.debug(s"Encoding STObject ${o.spaces4}")
     for {
       obj    <- prepareJsonObject(o, isSigning)
-      fields <- obj.traverse(TypeSerializers.encodeFieldAndValue(_, true, isSigning))
-    } yield EncodedSTObject(fields, isNested)
+      fields <- obj.traverse(TypeSerializers.encodeFieldAndValue(_, signingModeOn = isSigning))
+    } yield EncodedSTObject(fields, isTopLevel = false)
   }
 
   def decodeSTObject(v: List[UByte], info: FieldMetaData): Either[BCLibErr, (DecodedNestedField, List[UByte])] = {
-    val endOfSTObjectMarker = UByte(0xE1)
+    val stObjectEndByte = UByte(0xE1)
 
     // info is for the top level array
     @scala.annotation.tailrec
     def subfields(ub: List[UByte], acc: List[Decoded]): Either[BCLibErr, (List[Decoded], List[UByte])] = {
       ub match {
-        case h :: t if h === endOfSTObjectMarker => (acc.reverse, ub.drop(1)).asRight
-        case Nil                                 => BinCodecLibError("Badly Formed STObject").asLeft
+        case h :: t if h === stObjectEndByte => (acc.reverse, ub.drop(1)).asRight
+        case Nil                             => BinCodecLibError("Badly Formed STObject").asLeft
         case other => // The next bytes have another field
           TxBlobBuster.decodeNextField(ub) match {
             case Left(err)            => err.asLeft
@@ -84,10 +84,21 @@ trait STObjectCodec extends CodecUtils with JsonUtils {
 trait STArrayCodec extends CodecUtils with JsonUtils {
 
   /** Encodes each element of an array as an STObject.  */
-  def encodeSTArray(data: FieldData, isSigning: Boolean): Either[BinCodecLibError, EncodedSTArray] = {
+  def encodeSTArray(data: Json, isSigning: Boolean): Either[BinCodecLibError, EncodedSTArray] = {
+    scribe.debug(s"Encoding STArray ${data.spaces4}")
     for {
-      arr  <- json2array(data.json)
-      vals <- arr.traverse(j => STObjectCodec.encodeSTObject(j, isNested = false, isSigning = isSigning))
+      arr <- json2array(data)
+      // Meh, a nasty hack because we do not want an End of Object Marker on the top level object in the array
+      // Before I unwrapped, on basis that all STArrays are an array of oobject.
+      // Not sure why I changed. Could check if top level objects have just one field
+      //   _ = arr.forall(sj => sj.asObject.map(_.size)
+      vals <- arr.traverse { j =>
+        if (j.asObject.map(_.size) === Some(1))
+          STObjectCodec.encodeSTObject(j, isSigning = isSigning).map(_.copy(isTopLevel = true))
+        else {
+          BCJsonErr("Array wasnt all single field nested object", j).asLeft
+        }
+      }
     } yield EncodedSTArray(vals)
 
   }
@@ -120,10 +131,10 @@ trait Vector256Codec extends CodecUtils with JsonUtils {
     *
     * @return
     */
-  def encodeVector256(data: FieldData): Either[BinCodecLibError, EncodedVector256] = {
+  def encodeVector256(data: Json): Either[BinCodecLibError, EncodedVector256] = {
 
     for {
-      arr    <- json2array(data.json)
+      arr    <- json2array(data)
       hashes <- arr.traverse(HashHexCodecs.encodeHash256)
       // Pendantic as we know the length for each really.
       totalLen = hashes.map(_.value.rawBytes.length).sum