+++ modules/core/src/test/scala/com/odenzo/ripple/bincodec/testkit/OTestUtils.scala
@@ -2,7 +2,7 @@ package com.odenzo.ripple.bincodec.testkit
 
 import com.odenzo.ripple.bincodec.BinCodecLibError
 import com.odenzo.ripple.bincodec.ErrorOr.ErrorOr
-import io.circe.{Json, Decoder}
+import io.circe.{Json, Decoder, ACursor, JsonObject}
 import java.net.URL
 import scala.io.{Source, BufferedSource}
 
@@ -31,12 +31,44 @@ trait OTestUtils extends JsonUtils {
   }
 
   def loadJsonResource(path: String): Either[BinCodecLibError, Json] = {
-    BinCodecLibError.wrap(s"Getting Resource $path") {
+    BinCodecLibError.handlingM(s"Getting Resource $path") {
       val resource: URL          = getClass.getResource(path)
       val source: BufferedSource = Source.fromURL(resource)
       val data: String           = source.getLines().mkString("\n")
       JsonUtils.parseAsJson(data)
     }
+  }
+
+  /** Usage {{{
+    *
+    * }}}
+    */
+  def fieldNameChangeEx(name: String, newName: String)(in: JsonObject): JsonObject = {
+    // If missing existing name return JsonObject unchanges.
+    // If oldname == null then i guess will add newName : null
+    val updated: Option[JsonObject] = in(name)
+      .map(oldVal => in.add(newName, oldVal))
+      .map(jo => jo.remove(name))
+    updated.getOrElse(in)
+  }
+
+  def changeObjectField(oldName: String, newName: String): ACursor => ACursor = {
+    prepareJsonObject(fieldNameChangeEx(oldName, newName))
+  }
+
+  /** *
+    * {{{
+    *     val changer = fieldNameChangeEx("oldName","newName")
+    *     Decoder[A].prepare(prepareJsonObject(changer))
+    * }}}
+    *
+    * @param fn
+    * @param in
+    *
+    * @return
+    */
+  def prepareJsonObject(fn: JsonObject => JsonObject)(in: ACursor): ACursor = {
+    in.withFocus(json => json.mapObject(jobj => fn(jobj)))
   }
 
 }