+++ modules/core/src/test/scala/com/odenzo/ripple/bincodec/utils/JsonUtilsTest.scala
@@ -35,23 +35,4 @@ class JsonUtilsTest extends FunSuite with OTestSpec with JsonUtils {
     JsonUtils.parseAsJson(badJson).isLeft shouldBe true
   }
 
-  // This just adds the top level fields, creating duplicate fields on collision I think
-  // (x+y)+ z == x + (y+z)
-  // empty = JsonObject.empty
-  // empty + x = x+empty = x
-  // I think Circe supplies a JsonObject monoid (plus more I am not sure)
-  test("Monoid") {
-    val sum = for {
-      a <- parseAsJsonObject(aJson)
-      b <- parseAsJsonObject(bJson)
-
-    } yield a |+| b
-
-    val s = getOrLog(sum)
-    scribe.info(s"Sum  ${s.asJson.spaces4}")
-
-    s("b").value shouldEqual "2b".asJson
-    s.size shouldEqual 3
-  }
-
 }